+++ auth-lib/src/main/java/com/spotify/sdk/android/authentication/AuthenticationClient.java
@@ -329,6 +329,17 @@ public class AuthenticationClient {
     }
 
     /**
+     * Helper method to clear any Spotify and Facebook cookies from the WebView browser.
+     * <p>
+     * Calling this method will not affect the validity of the obtained access tokens and current Player session.
+     *
+     * @param context Android context.
+     */
+    public static void clearCookies(Context context) {
+        LoginDialog.clearCookies(context);
+    }
+
+    /**
      * Extracts {@link com.spotify.sdk.android.authentication.AuthenticationResponse}
      * from the LoginActivity result.
      *