+++ auth-lib/src/main/java/com/spotify/sdk/android/authentication/WebViewUtils.java
@@ -1,0 +1,49 @@
+package com.spotify.sdk.android.authentication;
+
+import android.content.Context;
+import android.webkit.CookieManager;
+import android.webkit.CookieSyncManager;
+
+/*
+ * The methods in this class come from com.facebook.internal.Utility class
+ * in Facebook SDK for Android v3.23.0
+ * *
+ * Public repository URL: https://github.com/facebook/facebook-android-sdk
+ * Commit ID: 6b9427d160daa746b84b769c16a9c9e4d9969936
+ * File path: facebook/src/com/facebook/internal/Utility.java
+ */
+class WebViewUtils {
+
+    static void clearFacebookCookies(Context context) {
+        // setCookie acts differently when trying to expire cookies between builds of Android that are using
+        // Chromium HTTP stack and those that are not. Using both of these domains to ensure it works on both.
+        clearCookiesForDomain(context, "facebook.com");
+        clearCookiesForDomain(context, ".facebook.com");
+        clearCookiesForDomain(context, "https://facebook.com");
+        clearCookiesForDomain(context, "https://.facebook.com");
+    }
+
+    static void clearCookiesForDomain(Context context, String domain) {
+        // This is to work around a bug where CookieManager may fail to instantiate if CookieSyncManager
+        // has never been created.
+        CookieSyncManager syncManager = CookieSyncManager.createInstance(context);
+        syncManager.sync();
+
+        CookieManager cookieManager = CookieManager.getInstance();
+
+        String cookies = cookieManager.getCookie(domain);
+        if (cookies == null) {
+            return;
+        }
+
+        String[] splitCookies = cookies.split(";");
+        for (String cookie : splitCookies) {
+            String[] cookieParts = cookie.split("=");
+            if (cookieParts.length > 0) {
+                String newCookie = cookieParts[0].trim() + "=;expires=Sat, 1 Jan 2000 00:00:01 UTC;";
+                cookieManager.setCookie(domain, newCookie);
+            }
+        }
+        cookieManager.removeExpiredCookie();
+    }
+}